refactor(cell): split missions/console along lifecycle + framework seams
Split two over-cap files (#529 tech-debt) into directories with
re-export discipline; pure code movement, no logic/signature changes.

missions.rs (922) -> missions/:
- mod.rs: doc + shared ON_* method-index constants + pub-use re-exports
- resend.rs: resend_missions
- lifecycle.rs: accept_mission, abandon_mission (+ accept/offer-guard tests)
- progression.rs: advance_step, complete_objective, complete_mission_direct

console/mod.rs (945) -> slimmed to declarations + re-exports; carved out:
- registry.rs: Spec/Target types + the static COMMANDS table
- dispatch.rs: handle_console_command, resolve_target, exec
- parse.rs: parse_i32 / parse_f32 / parse_bool

Behavior-preserving - re-export discipline - zero public-surface change.

Part of #529
